basic faction2 login

+++ app/src/main/java/com/example/mymusicapp/LoginActivity.java
@@ -3,6 +3,8 @@ package com.example.mymusicapp;
 import android.app.Activity;
 import android.content.Intent;
 import android.content.SharedPreferences;
+import android.database.Cursor;
+import android.database.sqlite.SQLiteDatabase;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
 import android.view.View;
@@ -26,6 +28,7 @@ public class LoginActivity extends Activity {
     private EditText editTextPassword;
     private CheckBox checkBoxRememberPassword;
     private RadioButton radioButtonAgree;
+    private DatabaseHelper databaseHelper;
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -78,24 +81,47 @@ public class LoginActivity extends Activity {
                     }
                     else{
                         //核对登录信息
-
-                        //信息准确且选择记住密码
-                        editor=sharedPreferences.edit();
-                        if(checkBoxRememberPassword.isChecked()){
-                            editor.putBoolean("remember_password",true);
-                            editor.putString("userName",userName);
-                            editor.putString("password",password);
+                        databaseHelper=new DatabaseHelper(LoginActivity.this,"SongsApp.db",null,1);
+                        SQLiteDatabase db=databaseHelper.getWritableDatabase();
+                        Boolean isTrue=new Boolean(false);
+                        String selection="userName = ?";
+                        String[] selectionArgs={userName};
+                        Cursor cursor=db.query("Users",null,selection,selectionArgs,null,null,null);
+                        //判断根据用户名筛选出的密码中是否有符合的
+                        if(cursor.moveToFirst()){
+                            do{
+                                String passwordInDB=cursor.getString(cursor.getColumnIndexOrThrow("password"));
+                                if(password.equals(passwordInDB)){
+                                    isTrue=true;
+                                }
+                            }while(cursor.moveToNext()&&isTrue.equals(false));
+                            if(isTrue.equals(true)){
+                                //信息准确且选择记住密码
+                                editor=sharedPreferences.edit();
+                                if(checkBoxRememberPassword.isChecked()){
+                                    editor.putBoolean("remember_password",true);
+                                    editor.putString("userName",userName);
+                                    editor.putString("password",password);
+                                }else{
+                                    editor.clear();
+                                }
+                                editor.apply();
+                                //成功登录系统
+                                Intent intent=new Intent(LoginActivity.this, MainActivity.class);
+                                startActivity(intent);
+                            }else {
+                                //遍历数据库中账户密码后没有符合的
+                                Toast.makeText(LoginActivity.this,"用户名或密码错误",Toast.LENGTH_SHORT).show();
+                            }
                         }else{
-                            editor.clear();
+                            //数据库里没有该用户
+                            Toast.makeText(LoginActivity.this,"账户信息不存在，请先注册",Toast.LENGTH_SHORT).show();
                         }
-                        editor.apply();
                     }
                 }else {
-                    Toast.makeText(LoginActivity.this,"请勾选隐私协议",Toast.LENGTH_SHORT).show();
+                    Toast.makeText(LoginActivity.this,"请勾选同意协议",Toast.LENGTH_SHORT).show();
                 }
             }
         });
-
-
     }
 }
